painter: move font related methods into font.go
